+++ euporie/core/kernel/jupyter.py
@@ -136,6 +136,16 @@ class JupyterKernel(BaseKernel):
         )
         self.client_lock = threading.Lock()
 
+        # Accessing the kernel spec causes `readline` to be imported, which causes the
+        # terminal to be set to cooked mode on MacOS when run not on the main thread.
+        # The import  process leading to this is:
+        #   jupyter_client -> ipykernel -> ipython.core.debugger -> pdb -> readline
+        # We deliverately access the kernelspec here in cooked mode, causing ptk to
+        # return the terminal to raw mode when done
+        if threading.current_thread() != threading.main_thread():
+            with self.kernel_tab.app.input.cooked_mode():
+                _ = self.km.kernel_spec
+
     def _set_living_status(self, alive: bool) -> None:
         """Set the life status of the kernel."""
         if not alive: